requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
docling
langchain-core
langchain-groq
python-docx
